+++ prpr/src/core/note.rs
@@ -6,14 +6,14 @@ pub use crate::{
 };
 use macroquad::prelude::*;
 
-const HOLD_PARTICLE_INTERVAL: f32 = 0.15;
-const FADEOUT_TIME: f32 = 0.16;
-const BAD_TIME: f32 = 0.5;
+const HOLD_PARTICLE_INTERVAL: f64 = 0.15;
+const FADEOUT_TIME: f64 = 0.16;
+const BAD_TIME: f64 = 0.5;
 
 #[derive(Clone, Debug)]
 pub enum NoteKind {
     Click,
-    Hold { end_time: f32, end_height: f32 },
+    Hold { end_time: f64, end_height: f64 },
     Flick,
     Drag,
 }
@@ -33,9 +33,9 @@ pub struct Note {
     pub object: Object,
     pub kind: NoteKind,
     pub hitsound: HitSound,
-    pub time: f32,
-    pub height: f32,
-    pub speed: f32,
+    pub time: f64,
+    pub height: f64,
+    pub speed: f64,
     pub color: Color,
     pub fx_color: Option<Color>,
     pub judge_area: f32,
@@ -50,8 +50,8 @@ pub struct Note {
 pub struct RenderConfig<'a> {
     pub settings: &'a ChartSettings,
     pub ctrl_obj: &'a mut CtrlObject,
-    pub line_height: f32,
-    pub appear_before: f32,
+    pub line_height: f64,
+    pub appear_before: f64,
     pub draw_below: bool,
     pub incline_sin: f32,
 }
@@ -138,11 +138,11 @@ impl Note {
         // && self.ctrl_obj.is_default()
     }
 
-    pub fn update(&mut self, res: &mut Resource, parent_rot: f32, parent_tr: &Matrix, ctrl_obj: &mut CtrlObject, line_height: f32) {
+    pub fn update(&mut self, res: &mut Resource, parent_rot: f32, parent_tr: &Matrix, ctrl_obj: &mut CtrlObject, line_height: f64) {
         self.object.set_time(res.time);
         if let Some(color) = if let JudgeStatus::Hold(perfect, at, ..) = &mut self.judge {
             if res.time > *at {
-                *at += HOLD_PARTICLE_INTERVAL / res.config.speed;
+                *at += HOLD_PARTICLE_INTERVAL / res.config.speed as f64;
                 Some(self.fx_color.unwrap_or_else(|| {
                     if *perfect {
                         res.res_pack.info.fx_perfect()
@@ -168,8 +168,8 @@ impl Note {
         // && self.ctrl_obj.dead()
     }
 
-    fn init_ctrl_obj(&self, ctrl_obj: &mut CtrlObject, line_height: f32) {
-        ctrl_obj.set_height((self.height - line_height + self.object.translation.1.now() / self.speed) * RPE_HEIGHT / 2.);
+    fn init_ctrl_obj(&self, ctrl_obj: &mut CtrlObject, line_height: f64) {
+        ctrl_obj.set_height((self.height - line_height + self.object.translation.1.now() as f64 / self.speed) * RPE_HEIGHT as f64 / 2.);
     }
 
     pub fn now_transform(&self, res: &Resource, ctrl_obj: &CtrlObject, base: f32, incline_sin: f32) -> Matrix {
@@ -215,10 +215,10 @@ impl Note {
             ..self.color
         };
         color.a *= res.alpha * ctrl_obj.alpha.now_opt().unwrap_or(1.);
-        let spd = self.speed * ctrl_obj.y.now_opt().unwrap_or(1.);
+        let spd = self.speed * ctrl_obj.y.now_opt().unwrap_or(1.) as f64;
 
-        let line_height = config.line_height / res.aspect_ratio * spd;
-        let height = self.height / res.aspect_ratio * spd;
+        let line_height = config.line_height / res.aspect_ratio as f64 * spd;
+        let height = self.height / res.aspect_ratio as f64 * spd;
 
         let base = height - line_height;
         let cover_base = if !config.settings.hold_partial_cover {
@@ -226,7 +226,7 @@ impl Note {
         } else {
             match self.kind {
                 NoteKind::Hold { end_time: _, end_height } => {
-                    let end_height = end_height / res.aspect_ratio * spd;
+                    let end_height = end_height / res.aspect_ratio as f64 * spd;
                     end_height - line_height
                 }
                 _ => height - line_height,
@@ -256,10 +256,10 @@ impl Note {
             let mut color = color;
             if !config.draw_below {
                 let alpha = (self.time - res.time).min(0.) / FADEOUT_TIME + 1.;
-                color.a *= if self.fake && res.time >= self.time { 0. } else { alpha };
+                color.a *= if self.fake && res.time >= self.time { 0. } else { alpha as f32 };
             }
-            color.a *= mod_alpha;
-            res.with_model(self.now_transform(res, ctrl_obj, base, config.incline_sin), |res| {
+            color.a *= mod_alpha as f32;
+            res.with_model(self.now_transform(res, ctrl_obj, base as f32, config.incline_sin), |res| {
                 draw_center(res, tex, order, scale, color);
             });
         };
@@ -281,12 +281,12 @@ impl Note {
                     if res.time >= end_time {
                         return;
                     }
-                    let end_height = end_height / res.aspect_ratio * spd;
-                    color.a *= mod_alpha;
+                    let end_height = end_height / res.aspect_ratio as f64 * spd;
+                    color.a *= mod_alpha as f32;
 
                     let h = if self.time <= res.time { line_height } else { height };
-                    let bottom = h - line_height;
-                    let top = end_height - line_height;
+                    let bottom = (h - line_height) as f32;
+                    let top = (end_height - line_height) as f32;
                     let tex = &style.hold;
                     let ratio = style.hold_ratio();
                     // body
@@ -367,7 +367,7 @@ impl Note {
 }
 
 pub struct BadNote {
-    pub time: f32,
+    pub time: f64,
     pub kind: NoteKind,
     pub matrix: Matrix,
 }
@@ -389,7 +389,7 @@ impl BadNote {
                 },
                 self.kind.order(),
                 res.note_width,
-                Color::new(0.423529, 0.262745, 0.262745, (self.time - res.time).max(-1.) / BAD_TIME + 1.),
+                Color::new(0.423529, 0.262745, 0.262745, ((self.time - res.time).max(-1.) / BAD_TIME + 1.) as f32),
             );
         });
         true
